+++ libIRDB/test/check_thunks.hpp
@@ -1,3 +1,23 @@
+/*
+ * Copyright (c) 2014 - Zephyr Software LLC
+ *
+ * This file may be used and modified for non-commercial purposes as long as
+ * all copyright, permission, and nonwarranty notices are preserved.
+ * Redistribution is prohibited without prior written consent from Zephyr
+ * Software.
+ *
+ * Please contact the authors for restrictions applying to commercial use.
+ *
+ * THIS SOURCE IS PROVIDED "AS IS" AND WITHOUT ANY EXPRESS OR IMPLIED
+ * WARRANTIES, INCLUDING, WITHOUT LIMITATION, THE IMPLIED WARRANTIES OF
+ * MERCHANTIBILITY AND FITNESS FOR A PARTICULAR PURPOSE.
+ *
+ * Author: Zephyr Software
+ * e-mail: jwd@zephyr-software.com
+ * URL   : http://www.zephyr-software.com/
+ *
+ */
+
 
 #ifndef check_thunks_hpp
 #define check_thunks_hpp
@@ -9,7 +29,7 @@ void find_all_module_starts(libIRDB::FileIR_t* firp, std::set<int>& thunk_bases)
 void check_for_thunks(libIRDB::FileIR_t* firp, const std::set<int>& thunk_bases);
 
 
-bool possible_target(int p, uintptr_t at=0);
+bool possible_target(uintptr_t p, uintptr_t at=0);
 
 
 #endif
